requirements.txt
discord.py>=2.3.2
aiohttp>=3.8.5
openai>=1.3.0
Pillow>=10.0.0
